AC-856: extract task queue, consultation, human feedback, and notebook mixins from SQLiteStore
Pure move of four method groups into sibling mixin modules, following the
existing SQLiteHubStoreMixin/SQLiteMonitorStoreMixin precedent (stub-method
pattern, no logic changes). Shrinks sqlite_store.py from 1169 to 797 lines.
